+++ src/lua/commands/exit_to_title.lua
@@ -3,10 +3,7 @@
 trx.console.register({
   name = "title",
   help = "console/cmd/title/help",
-  run = function(args)
-    if args ~= "" then
-      return trx.console.Result.BAD_INVOCATION
-    end
+  run = function()
     trx.game.exit_to_title()
   end,
 })