+++ src/lua/commands/inf_sprint.lua
@@ -10,15 +10,13 @@ local KEY = "debug.enable_endless_sprint"
 trx.console.register({
   name = "restless",
   help = "console/cmd/inf_sprint/help",
+  args = function(parser)
+    parser:positional("state", { type = "boolean", optional = true })
+  end,
   run = function(args)
-    local enable
-    if args == "" then
+    local enable = args.state
+    if enable == nil then
       enable = not trx.config.get(KEY)
-    else
-      enable = trx.strings.parse_bool(args)
-      if enable == nil then
-        return trx.console.Result.BAD_INVOCATION
-      end
     end
 
     trx.config.set(KEY, enable)